% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{anim_plot_obesity}
\alias{anim_plot_obesity}
\title{Create plot}
\usage{
anim_plot_obesity(country = "world", time = 20, width = 1600, height = 1000)
}
\arguments{
\item{country}{str | What country you want to plot or world}

\item{time}{int | How long}

\item{width}{int | Width resolution}

\item{height}{int | Height resolution}
}
\value{

}
\description{
Create plot
}
\examples{
anim_plot_obesity()
}
