% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_workflows.R
\name{auto_compare_baseline}
\alias{auto_compare_baseline}
\title{Creates an automated plot comparing different participant groups from the
Multisite study. An alternative approch to using this automatic function
would be to manually create a ggplot with the "geom_scatter_column()" function
and "theme_chewr()".}
\usage{
auto_compare_baseline(comparison, ...)
}
\arguments{
\item{comparison}{NA}
}
\value{
plot
}
\description{
Creates an automated plot comparing different participant groups from the
Multisite study. An alternative approch to using this automatic function
would be to manually create a ggplot with the "geom_scatter_column()" function
and "theme_chewr()".
}
\examples{
plot_compare_groups(data = redcap_data,
            filter = "baseline",
            y_aes = "pdff_liver_cirle_mean",
            order = c("control", "obese", "intervention"),
            plot_title = "Liver fat",
            plot_subtitle = "Amount of fat in liver",
            plot_xlab = "Groups",
            plot_ylab = "\% of liver fat")
}
