% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_workflows.R
\name{auto_end_table}
\alias{auto_end_table}
\title{Extract participant information and writes it to an easy to view table.}
\usage{
auto_end_table(partcipant_id)
}
\value{
Data table graphic
}
\description{
Extract participant information and writes it to an easy to view table.
}
\examples{
extract_participant_end_parameters(3030)
}
