% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autoplot_multisite_baseline}
\alias{autoplot_multisite_baseline}
\title{Creates plot comparing different participant groups from the Multisite study.}
\usage{
autoplot_multisite_baseline(comparison, ...)
}
\arguments{
\item{comparison}{str | Variable to be compared}

\item{...}{list | List of standard ggplot arguments}
}
\value{
ggplot2 object
}
\description{
An alternative approach to using this automatic function is to manually
create a ggplot with the "geom_scatter_column()" function and "theme_chewr()".
}
\examples{
autoplot_multisite_baseline(data = redcap_data,
            filter = "baseline",
            y_aes = "pdff_liver_cirle_mean",
            order = c("control", "obese", "intervention"),
            plot_title = "Liver fat",
            plot_subtitle = "Amount of fat in liver",
            plot_xlab = "Groups",
            plot_ylab = "\% of liver fat")
}
