% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process_folder.R
\name{batch_process_folder}
\alias{batch_process_folder}
\title{Batch process all gum images in a folder}
\usage{
batch_process_folder(input_dir = ".", output_dir = "plots", protocol = "chew")
}
\arguments{
\item{input_dir}{Folder containing image files}

\item{output_dir}{Folder to save plots}

\item{protocol}{Either "chew" or "fixed" (default = "chew")}
}
\value{
A data frame summarizing H_SD per image (per side)
}
\description{
Batch process all gum images in a folder
}
