% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{chewr_style}
\alias{chewr_style}
\title{Changes plot theme to chewr (multisite) style. Used in combination
with ggplot.}
\usage{
chewr_style(font = "Helvetica")
}
\arguments{
\item{font}{string | Font type}
}
\value{
ggplot theme
}
\description{
Changes plot theme to chewr (multisite) style. Used in combination
with ggplot.
}
\examples{
ggplot(data) +
   geom_scatter_column() +
   chewr_style()
}
