% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labka.R
\name{combine_redcap_labka}
\alias{combine_redcap_labka}
\title{Combine data from labka export with redcap data.}
\usage{
combine_redcap_labka(labka_data, redcap_data, identifier = FALSE)
}
\arguments{
\item{labka_data}{Tibble | Tibble of lakba data}

\item{redcap_data}{Tibble | Tibble of redcap data}

\item{identifier}{bool | Whether or not to include CPR number in the return. You must have CPR number in your data when combining data with labka data.}
}
\value{
tibble
}
\description{
Combine data from labka export with redcap data.
}
\note{
You need to have CPR number included in both the labka and redcap data to use this function (left join is based on CPR number and date).
}
\examples{
combine_redcap_labka(labka_data = labka_data,
                     redcap_data = redcap_data,
                     identifier = TRUE)
}
