% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{custom_fill}
\alias{custom_fill}
\title{Custom fill function. Will only fill if there only exists data in
enrollment arm 1.}
\usage{
custom_fill(data_point, column, data)
}
\arguments{
\item{data}{}
}
\description{
Custom fill function. Will only fill if there only exists data in
enrollment arm 1.
}
