% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hue_from_image .R
\name{extract_hue_from_image}
\alias{extract_hue_from_image}
\title{Extract Hue Values from Image Data}
\usage{
extract_hue_from_image(img_df)
}
\arguments{
\item{img_df}{A data frame of filtered RGB pixels with \code{c.1}, \code{c.2}, \code{c.3} columns.}
}
\value{
A numeric vector of hue values.
}
\description{
Extract Hue Values from Image Data
}
