% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_labs}
\alias{extract_labs}
\title{Extracts label names from list.}
\usage{
extract_labs(args, ylab)
}
\arguments{
\item{args}{... |}

\item{ylab}{}
}
\value{

}
\description{
Extracts label names from list.
}
\examples{
labels <- extract_labs(args, ylab = comparison)
}
