% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_analysis.R
\name{extract_participant_end_parameters}
\alias{extract_participant_end_parameters}
\title{Extract participant information and writes it to an easy to view table.}
\usage{
extract_participant_end_parameters(partcipant_id)
}
\value{
Data table graphic
}
\description{
Extract participant information and writes it to an easy to view table.
}
\examples{
extract_participant_end_parameters(3030)
}
