% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_theme}
\alias{extract_theme}
\title{Extracts theme arguments}
\usage{
extract_theme(args)
}
\arguments{
\item{args}{...}
}
\description{
Extracts theme arguments
}
\examples{
theme <- extract_theme(args)
}
