% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcount.R
\name{fcount}
\alias{fcount}
\title{Make a sorted frequeny table for afactor}
\usage{
fcount(x)
}
\arguments{
\item{factor}{}
}
\value{
A tibble
}
\description{
Make a sorted frequeny table for afactor
}
\examples{
fcount(iris$Species)
}
