% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_nafld}
\alias{filter_nafld}
\title{Filters data on the basis of participants in the control and obese group should not have NAFLD.}
\usage{
filter_nafld(dat, token, arg = NULL)
}
\arguments{
\item{dat}{tibble | Data. I}

\item{token}{numeric | API token from redcap}

\item{arg}{tibble | Liver fat column to filter on. Is NULL if column is missing.}
}
\value{
tibble
}
\description{
Filters data on the basis of participants in the control and obese group should not have NAFLD.
}
\examples{
filter_nafld(dat = dat, token = "dasdsadas2239193219312", arg = "pdff_liver_cirle_mean")
}
