% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_scatter_column}
\alias{geom_scatter_column}
\title{Custom ggplot2 geom that creates a scatter column plot. Good to use when
comparing groups.}
\usage{
geom_scatter_column(
  mapping = NULL,
  jitter.params = list(),
  meanbar.params = list(),
  errorbar.params = list(),
  ...
)
}
\value{
ggplot geom
}
\description{
Custom ggplot2 geom that creates a scatter column plot. Good to use when
comparing groups.
}
\examples{
ggplot(data) +
   geom_scatter_column()
}
