% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_test}
\alias{geom_test}
\title{Testing w. arguments}
\usage{
geom_test(
  ...,
  jitter.params = list(),
  meanbar.params = list(),
  errorbar.params = list()
)
}
\arguments{
\item{errorbar_params}{}
}
\value{

}
\description{
Testing w. arguments
}
