% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{label_text}
\alias{label_text}
\title{Sets axis text as angled or wrap.}
\usage{
label_text(axis_text = "angled")
}
\arguments{
\item{axis_text}{}
}
\description{
Sets axis text as angled or wrap.
}
\examples{
label_text()
}
