% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labka.R
\name{labka_read_data}
\alias{labka_read_data}
\title{Read Labka data and converts to tibble.}
\usage{
labka_read_data(path, file_type = "xlsx", identifier = FALSE)
}
\arguments{
\item{path}{string | File location of Labka data}

\item{file_type}{string | File type - Currently only supports xlsx}

\item{identifier}{bool | Whether or not to include CPR number in the return. You must have CPR number in your data when combining data with labka data.}
}
\value{
tibble
}
\description{
Read Labka data and converts to tibble.
}
\note{
If you wish to combine lakba with redcap data you must include CPR number (identifier = TRUE)
}
\examples{
read_labka_data(path = "~/Documents/Data (Local)/lakba-export_2021-06-01.xlsx",
                file_type = "xlsx",
                indetifier = TRUE)
}
