% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesoscale.R
\name{msd_read_table}
\alias{msd_read_table}
\title{Reads MSR table (CSV) data into a tibble.}
\usage{
msd_read_table(path)
}
\arguments{
\item{path}{str | Path to csv file}
}
\value{
tibble
}
\description{
Reads MSR table (CSV) data into a tibble.
}
\note{
For raw data, use '''read_msd_raw'''
}
\examples{
read_msd_table("data.csv")
}
