% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_if_na}
\alias{null_if_na}
\title{Returns NULL if provided value is NA}
\usage{
null_if_na(value)
}
\arguments{
\item{value}{str}
}
\value{
NULL or str
}
\description{
Returns NULL if provided value is NA
}
\examples{
vector <- c(1, 2, 3)

null_if_na(vector[1])
null_if_na(vector[10])
}
