% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filtered_image.R
\name{plot_filtered_image}
\alias{plot_filtered_image}
\title{Plot and save the filtered gum pixels from image data}
\usage{
plot_filtered_image(img_df, output_dir = "plots")
}
\arguments{
\item{img_df}{A data frame output from \code{process_gum_image()}}

\item{output_dir}{Directory to save the PNG plot}
}
\value{
Path to saved image
}
\description{
Plot and save the filtered gum pixels from image data
}
