% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_groups}
\alias{plot_groups}
\title{Creates an automated plot comparing different participant groups from the
Multisite study. An alternative approch to using this automatic function
would be to manually create a ggplot with the "geom_scatter_column()" function
and "chewr_style()".}
\usage{
plot_groups(
  data,
  filter,
  y_aes,
  order = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_xlab = NULL,
  plot_ylab = NULL
)
}
\arguments{
\item{data}{tibble | Redcap, labka, or a combination of both.}

\item{filter}{string | What parameter do you want to filter on. I.e. baseline.}

\item{y_aes}{string | What y value to look at}

\item{order}{vector | The x-axis order}

\item{title}{string | Main title}

\item{subtitle}{string | Sub title}

\item{xlab}{string | Label on x-axis}

\item{ylab}{string | Label on y-axis}
}
\value{
plot
}
\description{
Creates an automated plot comparing different participant groups from the
Multisite study. An alternative approch to using this automatic function
would be to manually create a ggplot with the "geom_scatter_column()" function
and "chewr_style()".
}
\examples{
plot_groups(data = redcap_data,
            filter = "baseline",
            y_aes = "pdff_liver_cirle_mean",
            order = c("control", "obese", "intervention"),
            plot_title = "Liver fat",
            plot_subtitle = "Amount of fat in liver",
            plot_xlab = "Groups",
            plot_ylab = "\% of liver fat")
}
