% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_multisite_weight_loss}
\alias{plot_multisite_weight_loss}
\title{Creates plot comparing different participant groups from the Multisite study.
This function automatically compares weight loss}
\usage{
plot_multisite_weight_loss(comparison, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Creates plot comparing different participant groups from the Multisite study.
This function automatically compares weight loss
}
