% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_ck18}
\alias{read_ck18}
\title{Read CK18 data and returns as tibble.}
\usage{
read_ck18(
 
    path = "/Users/andersaskeland/Documents/0 - Multisite/Storage (Local)/10 - Data/CK-18/CK18_results.xlsx"
)
}
\arguments{
\item{path}{str | Path to excel file with CK18 results}
}
\value{
tibble
}
\description{
Read CK18 data and returns as tibble.
}
\examples{
# Read CK18 data
df <- read_ck18()
}
