% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_dxa}
\alias{read_dxa}
\title{Reads DXA data and returns a tibble. TODO}
\usage{
read_dxa(
  path = "/Users/andersaskeland/Documents/1 - Data/multisite_dxa/multisite_DXA.xlsx"
)
}
\arguments{
\item{path}{str | Path to excel file with DXA results}
}
\description{
Reads DXA data and returns a tibble. TODO
}
\examples{
# Read DXA data
df <- read_dxa()
}
