% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_homa}
\alias{read_homa}
\title{Read HOMA-ir from CSV file. Homa IR is calculated using computer model from
HOMA2.}
\usage{
read_homa(path = "~/Documents/1 - Data/multisite_homa/homa_ir.csv")
}
\arguments{
\item{path}{str | Path to excel file with PIIINP results}
}
\description{
Read HOMA-ir from CSV file. Homa IR is calculated using computer model from
HOMA2.
}
