% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_insulin}
\alias{read_insulin}
\title{Reads insulin and C-peptide data}
\usage{
read_insulin(
  results =
    "/Users/andersaskeland/Documents/0 - Multisite/Storage (Local)/10 - Data/Insulin_c-peptide/insulin_c-peptide.csv",
  layout =
    "/Users/andersaskeland/Documents/0 - Multisite/Storage (Local)/10 - Data/Insulin_c-peptide/sample_layout.xlsx"
)
}
\arguments{
\item{results}{path | Path to file containing results & identifiers}

\item{layout}{path | Path to file containing group information}
}
\value{
tibble
}
\description{
Reads insulin and C-peptide data
}
\examples{
# Read insulin and c-peptide data
df <- read_insulin()
}
