% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_insulin_cpeptide}
\alias{read_insulin_cpeptide}
\title{Title}
\usage{
read_insulin_cpeptide(
 
    results = "/Users/andersaskeland/Documents/0 - Multisite/Storage (Local)/10 - Data/Insulin_c-peptide/insulin_c-peptide.csv",
 
    layout = "/Users/andersaskeland/Documents/0 - Multisite/Storage (Local)/10 - Data/Insulin_c-peptide/sample_layout.xlsx"
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
