% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_labka}
\alias{read_labka}
\title{Read Labka data and converts to tibble.}
\usage{
read_labka(path, identifier = FALSE)
}
\arguments{
\item{path}{string | File location of Labka data}

\item{identifier}{bool | Whether or not to include CPR number in the return.
One must have CPR number in your data when combining RedCap data with Labka data.}
}
\value{
tibble
}
\description{
Read Labka data and converts to tibble.
}
\note{
If you wish to combine Lakba with redcap data you must include CPR number (identifier = TRUE)
}
\examples{
# Read labka data and get tibble with identifier
labka_read(path = "~/Documents/Data (Local)/lakba-export.xlsx",
                indetifier = TRUE)
}
