% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_lihep}
\alias{read_lihep}
\title{Reads lithium heparin data manually entered.}
\usage{
read_lihep(path = "~/Documents/1 - Data/multisite_lihep/data.xlsx")
}
\arguments{
\item{path}{str | Path to data location}
}
\description{
Reads lithium heparin data manually entered.
}
