% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_p3np}
\alias{read_p3np}
\title{Title}
\usage{
read_p3np(
 
    path = "/Users/andersaskeland/Documents/0 - Multisite/Storage (Local)/10 - Data/PIIINP/results.xlsx"
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
