% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_allocate_automatic}
\alias{redcap_allocate_automatic}
\title{Automatically transfer participant from allocation group to supplied group}
\usage{
redcap_allocate_automatic(
  participant_id,
  url = "https://redcap.rn.dk/api/",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Automatically transfer participant from allocation group to supplied group
}
