% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_anonymize_data}
\alias{redcap_anonymize_data}
\title{Randomizes continuous data.}
\usage{
redcap_anonymize_data(df, remove = NULL)
}
\arguments{
\item{remove}{vector of variables to be completly removed.}
}
\value{

}
\description{
Randomizes continuous data.
}
