% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_codebook}
\alias{redcap_codebook}
\title{Extract avaliable data variables from a redcap projet.}
\usage{
redcap_codebook(token, url = "https://redcap.rn.dk/api/")
}
\arguments{
\item{token}{string | API token from redcap}

\item{url}{string | Link to API website}
}
\value{
tibble
}
\description{
Extract avaliable data variables from a redcap projet.
}
\examples{
redap_codebook(token = "dsadas8e218271732172313712",
               url = "https://redcap.rn.dk/api/")
}
