% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_export}
\alias{redcap_export}
\title{Extracts/exports data from redcap (using httr) and converts it to a tibble.
API token is collected using R studio api when function is run (Avoids
collecting API token in Rhistory).}
\usage{
redcap_export(
  fields = NULL,
  records = NULL,
  redcap_uri = "https://redcap.rn.dk/api/",
  identifier = FALSE,
  ...
)
}
\arguments{
\item{fields}{str or vec | Single field or vector of fields to export.}

\item{records}{num | Single or several records to read. Defaults to ALL.}

\item{redcap_uri}{str | URL to redcap API. Defaults to RN server.}

\item{identifier}{bool | Whether or not to include CPR number in export.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Extra arguments.}
}
\value{
tibble
}
\description{
Extracts/exports data from redcap (using httr) and converts it to a tibble.
API token is collected using R studio api when function is run (Avoids
collecting API token in Rhistory).
}
\examples{
# Export BMI and weight for participants 1001 and 1002. Is exported as tibble.
df <- redcap_read(fields = c("bmi", "weight"),
            records = c(1001, 1002))
}
