% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_export_participant_info}
\alias{redcap_export_participant_info}
\title{Extract participant information and writes it to an easy to view table.}
\usage{
redcap_export_participant_info(partcipant_id)
}
\arguments{
\item{partcipant_id}{int | Participant ID.}
}
\value{
tbl
}
\description{
Extract participant information and writes it to an easy to view table.
}
\examples{
# Generate table for participant 3030.
redcap_participant_info(3030)
}
