% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_import}
\alias{redcap_import}
\title{Imports data into RedCap.}
\usage{
redcap_import(
  df,
  import,
  redcap_uri = "https://redcap.rn.dk/api/",
  overwrite = "normal",
  ...
)
}
\arguments{
\item{df}{tibble | Data frame to import into redcap. Supports data read using
\itemize{
\item read_labka()
\item read_ck18()
\item read_dxa()
\item read_p3np()
\item read_homa()
\item read_insulin()
\item read_timp1()
}}

\item{import}{str | Selection of which data to import. Selections:
\itemize{
\item labka
\item ck18
\item dxa
\item p3np or piiinp
\item homa
\item insulin
\item timp1
\item lihep
}}

\item{redcap_uri}{str | URL to redcap API. Defaults to RN server.}

\item{overwrite}{str | Overwrite existing data on redcap or not. Selections:
\itemize{
\item normal
}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Extra arguments.}
}
\description{
Imports data into RedCap.
}
\examples{
# Read PIIINP data
df_p3np <- read_p3np()
redcap_import(df_p3np, import = "p3np")

# Read TIMP1 data
df_timp1 <- read_timp1()
redcap_import(df_timp1, import = "timp1")

# Read LABKA data
df_labka <- read_labka(path = "~/Documents/1 - Data/Multisite-Labka/lakba-export_2022-09-13.xlsx",
                      identifier = T)
redcap_import(df_labka, import = "labka")

# Read CK18
df_ck18 <- read_ck18()
redcap_import(df_ck18, import = "ck18")

# Read insulin
df_insulin <- read_insulin()
redcap_import(df_insulin, import = "insulin")
}
