% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_import_lakba}
\alias{redcap_import_lakba}
\title{Imports lakba data into redcap.}
\usage{
redcap_import_lakba(labka_df, redcap_uri = "https://redcap.rn.dk/api/", ...)
}
\arguments{
\item{...}{}
}
\description{
Imports lakba data into redcap.
}
