% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_import_t1}
\alias{redcap_import_t1}
\title{Import T1 fibrosis data into redcap}
\usage{
redcap_import_t1(file, redcap_uri = "https://redcap.rn.dk/api/", ...)
}
\arguments{
\item{...}{}
}
\description{
Import T1 fibrosis data into redcap
}
