% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_import_timp1}
\alias{redcap_import_timp1}
\title{Imports results from p3np analysis into redcap}
\usage{
redcap_import_timp1(timp1_df, redcap_uri = "https://redcap.rn.dk/api/", ...)
}
\arguments{
\item{...}{}
}
\description{
Imports results from p3np analysis into redcap
}
