% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_participant_info}
\alias{redcap_participant_info}
\title{Extract participant information and writes it to an easy to view table.}
\usage{
redcap_participant_info(partcipant_id)
}
\arguments{
\item{partcipant_id}{}
}
\value{
Data table graphic
}
\description{
Extract participant information and writes it to an easy to view table.
}
\examples{
table_information(3030)
}
