% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_read}
\alias{redcap_read}
\title{Extracts data from redcap (using httr) and writes it to a easy to use tibble.
API token is collected using R studio api when function is run.}
\usage{
redcap_read(
  fields = NULL,
  records = NULL,
  redcap_uri = "https://redcap.rn.dk/api/",
  identifier = FALSE,
  ...
)
}
\arguments{
\item{fields}{str | Single or several fields to read from redcap}

\item{records}{num | Single or several records to read. Defaults to ALL}

\item{redcap_uri}{str | Url to redcap API}

\item{identifier}{bool | Whether or not to include CPR number in the return.}

\item{...}{Extra arguments}
}
\value{
tibble
}
\description{
Extracts data from redcap (using httr) and writes it to a easy to use tibble.
API token is collected using R studio api when function is run.
}
\note{
If you wish to combine lakba with redcap data you must include CPR number (identifier = TRUE)
}
\examples{
redcap_read(fields = c("bmi", "weight"),
            records = c(1001, 1002))
}
