% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_svg_wide}
\alias{save_svg_wide}
\title{Simple save svg.}
\usage{
save_svg_wide(path, plot)
}
\arguments{
\item{path}{str | File path}

\item{plot}{plot | Plot object to be saved}
}
\description{
Simple save svg.
}
\examples{
save_svg_wide("~/document/plot.png", plot)
}
