% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_relative_change}
\alias{stat_relative_change}
\title{Calculate the relative change between different time points. For example
difference in BMI at baseline compared to month 1.}
\usage{
stat_relative_change(df, column)
}
\arguments{
\item{df}{df | Dataframe object}

\item{column}{obj | Column name (Without quotes)}
}
\value{
Dataframe object
}
\description{
Calculate the relative change between different time points. For example
difference in BMI at baseline compared to month 1.
}
\examples{
stat_relative_change(df, bmi)
}
