% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_hue.R
\name{summarize_hue}
\alias{summarize_hue}
\title{Summarize hue standard deviation per subject (with optional grouping)}
\usage{
summarize_hue(df, combine_sides = TRUE)
}
\arguments{
\item{df}{A data frame with columns Hue, Name, Chews, and Side}

\item{combine_sides}{Logical. If TRUE, collapses across Side (default = TRUE)}
}
\value{
A tibble summarizing H_SD per Name (and optionally Chews and Side)
}
\description{
Summarize hue standard deviation per subject (with optional grouping)
}
