% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_hue_by_side.R
\name{summarize_hue_by_side}
\alias{summarize_hue_by_side}
\title{Summarize hue standard deviation per subject and side}
\usage{
summarize_hue_by_side(df)
}
\arguments{
\item{df}{Data frame with Hue, Name, Chews, and Side}
}
\value{
Summary data frame with H_SD by Name, Chews, and Side
}
\description{
Summarize hue standard deviation per subject and side
}
