% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_information}
\alias{table_information}
\title{Extract participant information and writes it to an easy to view table.}
\usage{
table_information(partcipant_id)
}
\arguments{
\item{partcipant_id}{}
}
\value{
Data table graphic
}
\description{
Extract participant information and writes it to an easy to view table.
}
\examples{
table_information(3030)
}
