% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_chewr}
\alias{theme_chewr}
\title{Changes plot theme to chewr (multisite) style. Used in combination
with ggplot.}
\usage{
theme_chewr(font = "Helvetica")
}
\arguments{
\item{font}{string | Font type}
}
\value{
ggplot theme
}
\description{
Changes plot theme to chewr (multisite) style. Used in combination
with ggplot.
}
\examples{
ggplot(data) +
   geom_scatter_column() +
   theme_chewr()
}
