% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{who_query}
\alias{who_query}
\title{Queries WHO API database and cleans data.}
\usage{
who_query(query)
}
\arguments{
\item{query}{str | Data to query}
}
\value{
tibble
}
\description{
Queries WHO API database and cleans data.
}
\examples{
who_query("obesity")
}
