% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hsd_to_excel.R
\name{write_hsd_to_excel}
\alias{write_hsd_to_excel}
\title{Write masticatory performance summary to Excel}
\usage{
write_hsd_to_excel(
  df_masticatory_performance,
  filename = "masticatory_performance.xlsx"
)
}
\arguments{
\item{df_masticatory_performance}{A data frame of summarized hue values (including H_SD)}

\item{filename}{Output Excel file name}
}
\description{
Write masticatory performance summary to Excel
}
