% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_paired_column}
\alias{geom_paired_column}
\title{Custom ggplot2 geom that creates a Paried scatter column plot. Good to use
in paired analyses.}
\usage{
geom_paired_column(paired_variable, scale = 1, color = "#2b8cbe")
}
\arguments{
\item{color}{}
}
\value{

}
\description{
Custom ggplot2 geom that creates a Paried scatter column plot. Good to use
in paired analyses.
}
