% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{geom_scatter_column}
\alias{geom_scatter_column}
\title{Custom geom. Creates a scatter column plot. Can only be used as a geom
for a ggplot object.}
\usage{
geom_scatter_column()
}
\value{
ggplot geom
}
\description{
Custom geom. Creates a scatter column plot. Can only be used as a geom
for a ggplot object.
}
\examples{
ggplot(data) +
   geom_scatter_column()
}
