% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_p3np}
\alias{read_p3np}
\title{Read PIIINP data and returns as tibble.}
\usage{
read_p3np(
  path =
    "/Users/andersaskeland/Documents/0 - Multisite/Storage (Local)/10 - Data/PIIINP/results.xlsx"
)
}
\arguments{
\item{path}{str | Path to excel file with PIIINP results}
}
\value{
tibble
}
\description{
Read PIIINP data and returns as tibble.
}
\examples{
# Read PIIINP data
df <- read_p3np()
}
