% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{read_redcap}
\alias{read_redcap}
\title{Extracts data from redcap (using RCurl) and writes it to a easy to use tibble.
API token is colleced using R studio api when function is run.}
\usage{
read_redcap(
  columns = NULL,
  column_types = NULL,
  url = "https://redcap.rn.dk/api/",
  identifier = FALSE,
  filter = FALSE,
  ...
)
}
\arguments{
\item{columns}{vector | Can contain up to 20 different variables. Use function "redcap_codebook()" to view avaliable variables.}

\item{column_types}{vector | Define column data types. Is in the order as displayed on redcap.}

\item{url}{string | Link to API website}

\item{identifier}{bool | Whether or not to include CPR number in the return. You must have CPR number in your data when combining data with labka data.}

\item{filter}{str | Filter out given data. Currently supported: 1. NAFLD - Removes people with 5\% or more liver fat in the control and obese group}

\item{...}{Extra arguments}
}
\value{
tibble
}
\description{
Extracts data from redcap (using RCurl) and writes it to a easy to use tibble.
API token is colleced using R studio api when function is run.
}
\note{
If you wish to combine lakba with redcap data you must include CPR number (identifier = TRUE)
}
\examples{
read_redcap(columns = c("bmi", "weight"),
            url = "https://redcap.rn.dk/api/")
}
