% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_timp1}
\alias{read_timp1}
\title{Title}
\usage{
read_timp1(
 
    path = "/Users/andersaskeland/Documents/0 - Multisite/Storage (Local)/10 - Data/TIMP1/timp1_results.xlsx"
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
