% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_allocate_automatic}
\alias{redcap_allocate_automatic}
\title{Automatically transfer participant from allocation group to supplied group}
\usage{
redcap_allocate_automatic(
  participant_id,
  url = "https://redcap.rn.dk/api/",
  ...
)
}
\arguments{
\item{participant_id}{int | Participant you want to transfer}

\item{url}{str | Url for redcap server}

\item{...}{Extra parameters}
}
\description{
Automatically transfer participant from allocation group to supplied group
}
\examples{
redcap_allocate_automatic(3030)
}
