% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_codebook}
\alias{redcap_codebook}
\title{Extract avaliable data variables from a redcap projet.
R studio API collects API key.}
\usage{
redcap_codebook(url = "https://redcap.rn.dk/api/", ...)
}
\arguments{
\item{url}{string | Link to API website}
}
\value{
tibble
}
\description{
Extract avaliable data variables from a redcap projet.
R studio API collects API key.
}
\examples{
redap_codebook(token = "dsadas8e218271732172313712",
               url = "https://redcap.rn.dk/api/")
}
