% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_chewr}
\alias{theme_chewr}
\title{Custom theme for ggplot object.}
\usage{
theme_chewr(font = "Helvetica", scale = 1)
}
\arguments{
\item{font}{str | Font type}

\item{scale}{int | How large text should be}
}
\value{
ggplot theme
}
\description{
Custom theme for ggplot object.
}
\examples{
ggplot(data) +
   geom_scatter_column() +
   theme_chewr(scale = 2)
}
